rename sleep-rand-300.c to sleep-rand.c in sdr_transceiver_ft8
